requirements.txt
python-dotenv
langchain-community
langchain-core
langchain-text-splitters
langchain-chroma
langchain-openai